requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.23
python-dotenv==0.20.0
jinja2==3.1.2
python-multipart==0.0.6
better-profanity==0.7.0
